Fixing issue with keeping tracker in sync
I found while doing a rewrite of the Dishonored system into another 2d20 system that we had issues with player and GM staying in sync for Momentum and Chaos. The issue was that the local variables for momentum and chaos were only being set on initial load of the tracker and when the tracker was changed, it would change back to the value from when the user logged in, before changing back when it hit the setTimeout update. Now when there is a change to a value it will also update the other local variable. There might be a better way to fix this, but quick and dirty this did fix the issue I was seeing with my players.